Use the blockspring-cli to get the actual code
makes it easier to devlop locally and push new code
